Consolidate Demo and Layouts stories into Examples section
- Merge pptx/docx Demo + Layouts stories into a single Examples.stories.ts
- Rename xlsx Demo → Examples for consistency across the three packages
- Rename private sample sections from PrivateDemo → PrivateExamples (title only)
- Update smoke test story IDs from *-layouts--* to *-examples--*
