+++ packages/pptx/src/Examples.stories.ts
@@ -1,15 +1,34 @@
 import type { Meta, StoryObj } from '@storybook/html';
-import { DocxDocument } from './document';
+import { buildViewerUI } from './PptxViewer.stories';
+import { PptxPresentation } from './presentation';
 
-type Args = Record<string, never>;
+type DemoArgs = { width: number };
+type LayoutArgs = Record<string, never>;
 
-const SAMPLE_URL = `${import.meta.env.BASE_URL}docx/demo/sample-1.docx`;
+const SAMPLE_URL = `${import.meta.env.BASE_URL}pptx/demo/sample-1.pptx`;
 
-const meta: Meta<Args> = {
-  title: 'DocxViewer/Layouts',
+const meta: Meta<DemoArgs> = {
+  title: 'PptxViewer/Examples',
+  argTypes: {
+    width: {
+      control: { type: 'range', min: 400, max: 1600, step: 40 },
+      description: 'Canvas render width (px) — used by the Demo story',
+    },
+  },
+  args: { width: 960 },
 };
 export default meta;
-type Story = StoryObj<Args>;
+
+type DemoStory = StoryObj<DemoArgs>;
+type LayoutStory = StoryObj<LayoutArgs>;
+
+export const Demo: DemoStory = {
+  name: 'Demo — single viewer (demo.pptx)',
+  render(args) {
+    const { root } = buildViewerUI(args, SAMPLE_URL);
+    return root;
+  },
+};
 
 function makeStatus(root: HTMLElement): HTMLDivElement {
   const s = document.createElement('div');
@@ -19,16 +38,13 @@ function makeStatus(root: HTMLElement): HTMLDivElement {
   return s;
 }
 
-// ---------------------------------------------------------------------------
-// Story 1: ScrollView — all pages stacked vertically
-// ---------------------------------------------------------------------------
-export const ScrollView: Story = {
-  name: 'ScrollView — stack all pages',
+export const ScrollView: LayoutStory = {
+  name: 'ScrollView — stack all slides',
   render() {
     const root = document.createElement('div');
     root.style.cssText = 'font-family:sans-serif;padding:16px;';
     const heading = document.createElement('h3');
-    heading.textContent = 'ScrollView — scroll through every page';
+    heading.textContent = 'ScrollView — scroll through every slide';
     heading.style.cssText = 'margin:0 0 8px;font-size:14px;';
     root.appendChild(heading);
     const status = makeStatus(root);
@@ -38,19 +54,19 @@ export const ScrollView: Story = {
       'max-height:720px;overflow-y:auto;border:1px solid #ccc;background:#f5f5f5;padding:12px;';
     root.appendChild(scroller);
 
-    DocxDocument.load(SAMPLE_URL)
-      .then(async (doc) => {
-        status.textContent = `Rendering ${doc.pageCount} pages…`;
-        const widthPx = 700;
-        for (let i = 0; i < doc.pageCount; i++) {
+    PptxPresentation.load(SAMPLE_URL)
+      .then(async (pres) => {
+        status.textContent = `Rendering ${pres.slideCount} slides…`;
+        const widthPx = 800;
+        for (let i = 0; i < pres.slideCount; i++) {
           const canvas = document.createElement('canvas');
           canvas.style.cssText =
-            'display:block;width:100%;max-width:700px;margin:0 auto 12px;' +
+            'display:block;width:100%;max-width:800px;margin:0 auto 12px;' +
             'background:#fff;box-shadow:0 1px 3px rgba(0,0,0,0.2);';
           scroller.appendChild(canvas);
-          await doc.renderPage(canvas, i, { width: widthPx });
+          await pres.renderSlide(canvas, i, { width: widthPx });
         }
-        status.textContent = `Loaded ${doc.pageCount} pages`;
+        status.textContent = `Loaded ${pres.slideCount} slides`;
       })
       .catch((e: Error) => {
         status.textContent = `Error: ${e.message}`;
@@ -61,48 +77,45 @@ export const ScrollView: Story = {
   },
 };
 
-// ---------------------------------------------------------------------------
-// Story 2: ThumbnailGrid — all pages as a CSS-grid of small canvases
-// ---------------------------------------------------------------------------
-export const ThumbnailGrid: Story = {
-  name: 'ThumbnailGrid — overview of all pages',
+export const ThumbnailGrid: LayoutStory = {
+  name: 'ThumbnailGrid — overview of all slides',
   render() {
     const root = document.createElement('div');
     root.style.cssText = 'font-family:sans-serif;padding:16px;';
     const heading = document.createElement('h3');
-    heading.textContent = 'ThumbnailGrid — every page at a glance';
+    heading.textContent = 'ThumbnailGrid — every slide at a glance';
     heading.style.cssText = 'margin:0 0 8px;font-size:14px;';
     root.appendChild(heading);
     const status = makeStatus(root);
 
     const grid = document.createElement('div');
     grid.style.cssText =
-      'display:grid;grid-template-columns:repeat(auto-fill,minmax(160px,1fr));gap:16px;';
+      'display:grid;grid-template-columns:repeat(auto-fill,minmax(240px,1fr));gap:16px;';
     root.appendChild(grid);
 
-    DocxDocument.load(SAMPLE_URL)
-      .then(async (doc) => {
-        status.textContent = `Rendering ${doc.pageCount} thumbnails…`;
-        const thumbWidth = 160;
-        for (let i = 0; i < doc.pageCount; i++) {
+    PptxPresentation.load(SAMPLE_URL)
+      .then(async (pres) => {
+        status.textContent = `Rendering ${pres.slideCount} thumbnails…`;
+        const thumbWidth = 240;
+        for (let i = 0; i < pres.slideCount; i++) {
           const cell = document.createElement('div');
           cell.style.cssText = 'display:flex;flex-direction:column;align-items:center;cursor:pointer;';
           const canvas = document.createElement('canvas');
           canvas.style.cssText =
-            'display:block;width:100%;max-width:160px;background:#fff;' +
+            'display:block;width:100%;max-width:240px;background:#fff;' +
             'box-shadow:0 1px 3px rgba(0,0,0,0.2);';
           const caption = document.createElement('div');
-          caption.textContent = `Page ${i + 1}`;
+          caption.textContent = `Slide ${i + 1}`;
           caption.style.cssText = 'font-size:12px;color:#444;margin-top:4px;';
           cell.append(canvas, caption);
           const idx = i;
           cell.addEventListener('click', () => {
-            console.log(`[docx ThumbnailGrid] clicked page ${idx + 1}`);
+            console.log(`[pptx ThumbnailGrid] clicked slide ${idx + 1}`);
           });
           grid.appendChild(cell);
-          await doc.renderPage(canvas, i, { width: thumbWidth });
+          await pres.renderSlide(canvas, i, { width: thumbWidth });
         }
-        status.textContent = `Loaded ${doc.pageCount} pages`;
+        status.textContent = `Loaded ${pres.slideCount} slides`;
       })
       .catch((e: Error) => {
         status.textContent = `Error: ${e.message}`;
@@ -113,10 +126,7 @@ export const ThumbnailGrid: Story = {
   },
 };
 
-// ---------------------------------------------------------------------------
-// Story 3: MasterDetail — scrollable thumbnails + a large preview canvas
-// ---------------------------------------------------------------------------
-export const MasterDetail: Story = {
+export const MasterDetail: LayoutStory = {
   name: 'MasterDetail — thumbnails + large preview',
   render() {
     const root = document.createElement('div');
@@ -133,7 +143,7 @@ export const MasterDetail: Story = {
 
     const thumbCol = document.createElement('div');
     thumbCol.style.cssText =
-      'flex:0 0 200px;overflow-y:auto;border:1px solid #ccc;background:#f5f5f5;padding:8px;' +
+      'flex:0 0 240px;overflow-y:auto;border:1px solid #ccc;background:#f5f5f5;padding:8px;' +
       'display:flex;flex-direction:column;gap:10px;';
     const detailCol = document.createElement('div');
     detailCol.style.cssText =
@@ -144,43 +154,43 @@ export const MasterDetail: Story = {
     detailCol.appendChild(detailCanvas);
     layout.append(thumbCol, detailCol);
 
-    DocxDocument.load(SAMPLE_URL)
-      .then(async (doc) => {
-        status.textContent = `Rendering ${doc.pageCount} thumbnails…`;
+    PptxPresentation.load(SAMPLE_URL)
+      .then(async (pres) => {
+        status.textContent = `Rendering ${pres.slideCount} thumbnails…`;
         const thumbEntries: HTMLDivElement[] = [];
 
         const detailWidth = () => detailCol.clientWidth - 24;
 
-        const selectPage = async (i: number) => {
+        const selectSlide = async (i: number) => {
           for (let k = 0; k < thumbEntries.length; k++) {
             thumbEntries[k].style.outline = k === i ? '2px solid #0366d6' : 'none';
           }
-          await doc.renderPage(detailCanvas, i, { width: Math.max(320, detailWidth()) });
+          await pres.renderSlide(detailCanvas, i, { width: Math.max(320, detailWidth()) });
         };
 
-        for (let i = 0; i < doc.pageCount; i++) {
+        for (let i = 0; i < pres.slideCount; i++) {
           const cell = document.createElement('div');
           cell.style.cssText = 'display:flex;flex-direction:column;align-items:center;cursor:pointer;padding:4px;';
           const canvas = document.createElement('canvas');
           canvas.style.cssText =
-            'display:block;width:100%;max-width:180px;background:#fff;box-shadow:0 1px 3px rgba(0,0,0,0.2);';
+            'display:block;width:100%;max-width:220px;background:#fff;box-shadow:0 1px 3px rgba(0,0,0,0.2);';
           const caption = document.createElement('div');
-          caption.textContent = `Page ${i + 1}`;
+          caption.textContent = `Slide ${i + 1}`;
           caption.style.cssText = 'font-size:12px;color:#444;margin-top:4px;';
           cell.append(canvas, caption);
           const idx = i;
           cell.addEventListener('click', () => {
-            selectPage(idx).catch((e: Error) => {
+            selectSlide(idx).catch((e: Error) => {
               status.textContent = `Render error: ${e.message}`;
             });
           });
           thumbCol.appendChild(cell);
           thumbEntries.push(cell);
-          await doc.renderPage(canvas, i, { width: 180 });
+          await pres.renderSlide(canvas, i, { width: 220 });
         }
 
-        await selectPage(0);
-        status.textContent = `Loaded ${doc.pageCount} pages`;
+        await selectSlide(0);
+        status.textContent = `Loaded ${pres.slideCount} slides`;
       })
       .catch((e: Error) => {
         status.textContent = `Error: ${e.message}`;
