+++ packages/xlsx/src/Examples.stories.ts
@@ -4,7 +4,7 @@ import { buildViewerUI } from './XlsxViewer.stories';
 type Args = { scale: number };
 
 const meta: Meta<Args> = {
-  title: 'XlsxViewer/Demo',
+  title: 'XlsxViewer/Examples',
   argTypes: {
     scale: {
       control: { type: 'range', min: 0.25, max: 2, step: 0.05 },
@@ -17,7 +17,7 @@ export default meta;
 type Story = StoryObj<Args>;
 
 export const Demo: Story = {
-  name: 'demo.xlsx',
+  name: 'Demo — single viewer (demo.xlsx)',
   render(args) {
     const { root } = buildViewerUI(args, `${import.meta.env.BASE_URL}xlsx/demo/sample-1.xlsx`);
     return root;
